add custom decoding to support Runtime type
Update the movie input stuct to use data.Runtime instead of int.

Implement a UnmarshalJSON() method on the Runtime type so that it satisfies the json.Unmarshaler interface.

IMPORTANT: Because UnmarshalJSON() needs to modify the receiver (our Runtime type), we must use a pointer receiver for this to work correctly.

Otherwise, we will only be modifying a copy (which is then discarded when this method returns)