Validating JSON Input
custom Validator type which contains a map of errors. The Validator type provides a Check() method for conditionally adding errors to the map, and a Valid() method which returns whether the errors map is empty or not. We’ve also added PermittedValue(), Matches() and Unique() functions to help us perform some specific validation checks.

Add validation checks to createMovie Handler